Implemented setting maximum size for content to be cached
When a zip file containing a lot of small files, less then 512k, the
ZipFile will crash due to caching all of the content in memory. Fixed this
by adding options to be passed along the zip file object to se the memory
threshold. When set to 0 it will not cache file contents and not fail the
memory limit.